Add a load result interface for use in new load_rules methods
Define a falco_load_result abstract class for use in new load_rules
methods. It's abstract so the implementation details in
rule_loader/rule_reader can be hidden from someone who wants to use
the API to load rules and work with a result.

The class defines a set of error codes/warning codes and has static
methods to get a short and long description of each error/warning.

There are virtual methods to access the important parts of a result:
 - successful or not
 - required engine version (to provide parity with older methods that
   returned this.
 - a summary of the result, suitable for short log entries.
 - a string representation of the result, suitable for display to
   users.
 - a json representation of the result, suitable for automated
   parsing/interpretation later

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>